--- a/util/Cargo.toml
+++ b/util/Cargo.toml
@@ -39,6 +39,7 @@ target_info = "0.1"
 [features]
 default = []
 dev = ["clippy"]
+x64asm = []
 
 [build-dependencies]
 vergen = "*"