--- a/util/src/uint.rs
+++ b/util/src/uint.rs
@@ -222,8 +222,7 @@ macro_rules! uint_overflowing_mul {
 				adc $$0, %rdx
 				or %rdx, %rcx
 
-				cmpq $$0, %rcx
-				jne 2f
+				jrcxz 2f
 
 				mov $8, %rax
 				cmpq $$0, %rax
@@ -233,6 +232,8 @@ macro_rules! uint_overflowing_mul {
 				cmpq $$0, %rax
 				sete %dl
 				or %dl, %cl
+
+				jrcxz 2f
 
 				mov $3, %rax
 				cmpq $$0, %rax