@NikVolf 2016-02-25
just good old asm :)