@arkpar 2016-02-25
What about U512? That's even more of a bottleneck currently.
I think we don't actually need multiplication of 512 -bit uints, We can get away with a custom function that multiplies two U256s into a U512.