@arkpar 2016-02-25
`Sub::sub` is still not using the asm version?