@NikVolf 2016-02-25
@arkpar `sub` uses `overflowing_add`
so it should become better as well

@arkpar 
only u256 in this pr
u512 comes next

@arkpar 
other issues resolved