@arkpar 2016-02-26
build warnings https://travis-ci.org/ethcore/parity/jobs/111836764#L824