@debris 2016-02-26
You can also move assembly code to separate `.s` files and compile it with gcc. It would make optimization available also on rust beta/stable.