@arkpar 2016-02-26
@marek We want the calls to be inlined. Also less less problems with windows later