@NikVolf 2016-02-26
removed warning (cause it's macros expansion, had to ignore it)