@NikVolf 2016-02-26
@gavofyork 
yeah right
i have plenty of tests here: https://github.com/NikVolf/asm-fun/blob/master/src/lib.rs
will migrate it