@NikVolf 2016-02-26
@debris indeed, Arkady right, inlining gives 2x boost to performance for add/sub